Create class for MobileContentRendererView and MobileContentRendererViewModel which subclass MobileContentPagesView and MobileContentPagesViewModel
